real_estate_scraper/requirements.txt
apache-airflow==2.10.5
psycopg2_binary==2.9.10
python-dotenv==1.1.0
Scrapy==2.11.2
